elasticsearch/requirements.txt
django == 4.2.5
djangorestframework == 3.14.0
langchain
langsmith
langchain_google_genai
langchain-elasticsearch
elasticsearch
rank_bm25
django-cors-headers
langchain_community